upgrade ws chat broker dependencies
