+++ Cargo.toml
@@ -46,7 +46,7 @@ members = [
    "web-cors/backend",
    "websocket",
    "websocket-chat",
-#   "websocket-chat-broker",
+   "websocket-chat-broker",
    "websocket-tcp-chat",
    "docker_sample",
 ]
