+++ websocket-chat-broker/Cargo.toml
@@ -1,7 +1,10 @@
 [package]
-name = "websocket-broker-example"
+name = "websocket-chat-broker"
 version = "0.1.0"
-authors = ["Chris Ricketts <chris.ricketts@steribar.com>"]
+authors = [
+  "Chris Ricketts <chris.ricketts@steribar.com>",
+  "Rob Ede <robjtede@icloud.com>",
+]
 edition = "2018"
 
 [[bin]]
@@ -9,12 +12,13 @@ name = "server"
 path = "src/main.rs"
 
 [dependencies]
-rand = "0.6"
-futures = "0.1.24"
-actix = "0.8.2"
-actix-web = "1.0"
-actix-files = "0.1"
-actix-web-actors = "1.0"
-actix-broker = "0.2.0"
-log = "0.4.5"
-simple_logger = "0.5.0"
+actix = "0.9"
+actix-broker = "0.3.0"
+actix-files = "0.2"
+actix-rt = "1.1"
+actix-web = "2.0"
+actix-web-actors = "2.0"
+env_logger = "0.7"
+futures = "0.3"
+log = "0.4"
+rand = "0.7"
